+++ Runtime/StateMachine/Core/IState.cs
@@ -1,0 +1,9 @@
+namespace NekoFlow
+{
+    public interface IState
+    {
+        void OnEnter();
+        void OnTick(float deltaTime);
+        void OnExit();
+    }
+}