+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [2.0.0] - 2025-12-14
+
+### State Machine & Flow Behaviour
+
+- Added ability to get time spent in the current state via StateBehaviour.GetTimeInCurrentState().
+- Improved transition management to ensure current transitions are updated correctly when adding new transitions.
+- Enhanced StateBehaviourInspector to provide clearer runtime information display.
+- Added SimpleFlow and BranchFlow.
+
 ## [1.0.0] - 2025-09-20
 
 ### Flow Machine